// src/types/scholarship.ts
export enum ScholarshipType {
  ACADEMIC = 'ACADEMIC',
  NEED_BASED = 'NEED_BASED',
  ORPHAN = 'ORPHAN',
  SPECIAL_NEEDS = 'SPECIAL_NEEDS',
  REFUGEE = 'REFUGEE',
  DISASTER_VICTIM = 'DISASTER_VICTIM',
  TALENT = 'TALENT',
}

export enum ApplicationStatus {
  DRAFT = 'DRAFT',
  SUBMITTED = 'SUBMITTED',
  UNDER_REVIEW = 'UNDER_REVIEW',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  WAITLIST = 'WAITLIST',
  WITHDRAWN = 'WITHDRAWN',
  CANCELLED = 'CANCELLED',
}